requirements.txt
pika==1.3.2
python-dotenv==0.20.0
lorem==0.1.1